% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPhylopic.R
\name{addPhylopic}
\alias{addPhylopic}
\title{Add transparency and/or change color of a picture}
\usage{
addPhylopic(
  img,
  x = NULL,
  y = NULL,
  ysize = NULL,
  alpha = 0.2,
  color = NULL,
  AR = 1
)
}
\arguments{
\item{img}{\link{array} a picture loaded by \code{png::readPNG()}.}

\item{x}{\link{numeric} the x coordinate of the center of the picture.}

\item{y}{\link{numeric} the y coordinate of the center of the picture.}

\item{ysize}{\link{numeric} a facteur to resize picture dimensions.}

\item{alpha}{\link{numeric} transparency factor between 0 (transparent) and 1 (opaque).}

\item{color}{\link{array} a color name (e.g. "red", "steelblue") or an hexadecimal value (e.g. "#ff0000")}

\item{AR}{\link{array} a magnification factor (zoom).}
}
\value{
This function returns a picture matrix with colors and/or transparency altered.
This matrix can be plotted with the function \code{graphics::rasterImage()} (see examples).
}
\description{
This function recolours and plots a picture. Not perfect, needs improvement.
}
\examples{

# No example.
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.com}
}
