% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/__findHull.R
\name{findHull}
\alias{findHull}
\title{Get corners of convex hull of a set of points}
\usage{
findHull(data)
}
\arguments{
\item{data}{[data.frame] a data frame with at least two columns named x and y.}
}
\value{
This function returns a susbet of the initial data frame where each row is
a corner of the convex hull.
}
\description{
This function gets corners coordinates of the convex hull of a set of points.
}
\examples{

x   <- rnorm(100)
y   <- rnorm(100)
mat <- data.frame(x, y)

dat <- findHull(mat)

dat <- rbind(dat, dat[1, ])

plot(mat, pch = 19)
polygon(dat[ , 1], dat[ , 2], border = "red")
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.com}

Nicolas LOISEAU, \email{nicolas.loiseau1@gmail.com}
}
