% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/__plotRVB.R
\name{plotRVB}
\alias{plotRVB}
\title{Apply a color gradient to a raster}
\usage{
plotRVB(x = NULL, zlim = NULL, type = NULL, palette = "Spectral",
  reverse = TRUE, alpha = 255, bgalpha = 0, colNA = NA,
  axes = FALSE, xlab = "", ylab = "", add = FALSE, zmax = NULL)
}
\arguments{
\item{x}{[RasterLayer] a RasterLayer to plot.}

\item{zlim}{[numeric] a vector of length 2. Range of values to plot.}

\item{type}{[string] not implemented.}

\item{palette}{[string] not well implemented.}

\item{reverse}{[boolean] not well implemented.}

\item{alpha}{[integer] transparency factor between 0 (transparent) and 255 (opaque).}

\item{bgalpha}{[integer] transparency factor for NA values between 0 (transparent) and 255 (opaque).}

\item{colNA}{[string] color for the background (i.e. NA values).}

\item{axes}{[boolean] If TRUE, axes are added.}

\item{xlab}{[string] x-axis label.}

\item{ylab}{[string] y-axis label.}

\item{add}{[boolean] If TRUE, raster is added the active graphical device.}

\item{zmax}{[numeric] not implemented.}
}
\value{
This function returns a 3-bands raster with the first layer is for the Red,
the second for the Green and the third for the Blue. Each non NA cell contains
the quantity of a specific color (R, G or B).
}
\description{
This function applies a color gradient to a raster where color is defined by the user.
}
\examples{

# No example.
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.com}
}
