% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRVB2.R
\name{plotRVB2}
\alias{plotRVB2}
\title{Apply a color gradient to a raster}
\usage{
plotRVB2(
  x,
  n_classes = NULL,
  breaks = NULL,
  palettes = NULL,
  alpha = 255,
  bgalpha = 0,
  colNA = NA,
  axes = FALSE,
  xlab = "",
  ylab = "",
  add = FALSE
)
}
\arguments{
\item{x}{\link{RasterLayer} a RasterLayer to plot.}

\item{n_classes}{\link{...} ...}

\item{breaks}{\link{...} ...}

\item{palettes}{\link{...} ...}

\item{alpha}{\link{integer} transparency factor between 0 (transparent) and 255 (opaque).}

\item{bgalpha}{\link{integer} transparency factor for NA values between 0 (transparent) and 255 (opaque).}

\item{colNA}{\link{string} color for the background (i.e. NA values).}

\item{axes}{\link{boolean} If TRUE, axes are added.}

\item{xlab}{\link{string} x-axis label.}

\item{ylab}{\link{string} y-axis label.}

\item{add}{\link{boolean} If TRUE, raster is added the active graphical device.}
}
\value{
This function returns a 3-bands raster with the first layer is for the Red,
the second for the Green and the third for the Blue. Each non NA cell contains
the quantity of a specific color (R, G or B).
}
\description{
This function applies a color gradient to a raster where color is defined by the user.
}
\examples{

# No example.
 -------------------------------------------------------------------------   @AddLegendLabel
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.com}
}
