% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recolorPhylopic.R
\name{recolorPhylopic}
\alias{recolorPhylopic}
\title{Add transparency and/or change color of a picture}
\usage{
recolorPhylopic(img, alpha = 0.2, color = NULL)
}
\arguments{
\item{img}{\link{array} a picture loaded by \code{png::readPNG()}.}

\item{alpha}{\link{numeric} transparency factor between 0 (transparent) and 1 (opaque).}

\item{color}{\link{array} a color name (e.g. "red", "steelblue") or an hexadecimal value (e.g. "#ff0000")}
}
\value{
This function returns a picture matrix with colors and/or transparency altered.
This matrix can be plotted with the function \code{graphics::rasterImage()} (see examples).
}
\description{
This function adds transparency and/or change color of a picture (array).
}
\examples{

# Download Rlogo in png format
download.file(
  url      = "https://www.r-project.org/logo/Rlogo.png",
  destfile = "Rlogo.png"
)

# Import Rlogo (require `png` package)
rlogo <- png::readPNG("Rlogo.png")

# Get image dimensions
xmax <- dim(rlogo)[2]
ymax <- dim(rlogo)[1]

# Remove figure margins
o_par <- par()
par(mar = rep(0, 4))


### ORIGINAL IMAGE   ------


# Empty plot
xmin <- 0
ymin <- xmin

plot(
  x    = xmin,
  x    = ymin,
  xlim = c(xmin, xmax), # X-axis extent (number of image rows)
  ylim = c(ymin, ymax), # Y-axis extent (number of image cols)
  asp  = 1,             # Preserve x-axis/y-axis ratio
  ann  = FALSE,         # Remove figure annotations
  axes = FALSE,         # Remove figure axes
  type = "n"            # Remove figure data
)

# Add Rlogo
graphics::rasterImage(
  image   = rlogo,
  xleft   = xmin,
  ybottom = ymin,
  xright  = xmax,
  ytop    = ymax,
  angle   = 0
)


### IMAGE WITH TRANSPARENCY   ------

# Empty plot
plot(
  x    = 0,
  xlim = c(0, xmax),
  ylim = c(0, ymax),
  asp  = 1,
  ann  = FALSE,
  axes = FALSE,
  type = "n"
)

# Add transparency
rlogo_alpha <- recolorPhylopic(rlogo, alpha = 0.25, color = NULL)

# Add Rlogo
graphics::rasterImage(
  image   = rlogo_alpha,
  xleft   = 0,
  ybottom = 0,
  xright  = xmax,
  ytop    = ymax,
  angle   = 0
)


### MONOCHROMATIC IMAGE   ------

# Empty plot
plot(
  x    = 0,
  xlim = c(0, xmax),
  ylim = c(0, ymax),
  asp  = 1,
  ann  = FALSE,
  axes = FALSE,
  type = "n"
)

# Add transparency
rlogo_blue <- recolorPhylopic(rlogo, alpha = 1, color = "steelblue")

# Add Rlogo
graphics::rasterImage(
  image   = rlogo_blue,
  xleft   = 0,
  ybottom = 0,
  xright  = xmax,
  ytop    = ymax,
  angle   = 0
)

# Delete downloaded Rlogo
file.remove("Rlogo.png")

# Reset default par()
par(o_par)

}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.com}

Nicolas LOISEAU, \email{nicolas.loiseau1@gmail.com}
}
