% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGraticules.R
\name{addGraticules}
\alias{addGraticules}
\title{Add graticules (parallels and meridians) to a map}
\usage{
addGraticules(
  prj = "+proj=cea +lon_0=0 +lat_ts=30 +x_0=0 +y_0=0 +ellps=WGS84 +units=m +no_defs",
  parallels = seq(-60, 60, by = 30),
  meridians = seq(-160, 160, by = 40),
  line.color = "#aaaaaa",
  line.size = 0.5,
  line.type = 1,
  add = FALSE
)
}
\arguments{
\item{prj}{\link{string} the CRS of the map in the PROJ4 standard.}

\item{parallels}{\link{numeric} vector of latitudes to add parallels.}

\item{meridians}{\link{numeric} vector of longitudes to add meridians.}

\item{line.color}{\link{string} color of graticules.}

\item{line.size}{\link{numeric} width of graticules.}

\item{line.type}{\link{integer} type of graticules.}

\item{add}{\link{boolean} If TRUE, graticules are added to the map.}
}
\value{
This function returns a 2-elements list.

A first data frame for the meridians with 4 columns:
\itemize{
\item x: the x-coordinates of meridians
\item y: the y-coordinates of meridians
\item label: the labels of meridians (in degrees)
\item direction: the direction, i.e. W(est) or E(ast)
}

A second data frame for the parallels with 4 columns:
\itemize{
\item x: the x-coordinates of parallels
\item y: the y-coordinates of parallels
\item label: the labels of parallels (in degrees)
\item direction: the direction, i.e. S(outh) or N(orth)
}
}
\description{
This function adds graticules (parallels and meridians) to a map and returns
axes labels and coordinates.
}
\examples{

library(maps)

maps::map()
addGraticules(
  prj = "+init=epsg:4326",
  add = TRUE
)

## [[1]]
##      x      y label direction
## 1 -160 -90.00   160         W
## 2 -120 -89.95   120         W
## 3  -80 -89.90    80         W
## 4  -40 -89.85    40         W
## 5    0 -89.80     0
## 6   40 -89.75    40         E
## 7   80 -89.70    80         E
## 8  120 -89.65   120         E
## 9  160 -89.60   160         E

## [[2]]
##        x   y label direction
## 1 -180.0 -60    60         S
## 2 -179.9 -30    30         S
## 3 -179.8   0     0
## 4 -179.7  30    30         N
## 5 -179.6  60    60         N

}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.com}
}
