% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/__addInset.R
\name{addInset}
\alias{addInset}
\title{Add inset map to an existing map}
\usage{
addInset(x = NULL, region = NULL, where = NULL, zoom = 2,
  title = NULL, case = 1, yat = 0.1)
}
\arguments{
\item{x}{[raster] a RasterLayer to crop and add into an inset.}

\item{region}{[vector] the extent of the crop box (xmn, xmx, ymn, ymx).}

\item{where}{[vector] the coordinates of left-bottom corner of the inset.}

\item{zoom}{[numeric] a zoom factor.}

\item{title}{[string] the inset title.}

\item{case}{[integer] predefined case.}

\item{yat}{[numeric] a factor to correct title position.}
}
\value{
For an internal use only.
}
\description{
For an internal use only.
}
\examples{

# No example.

}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.com}
}
